migrated rosbag opening function + helpers to separate file; changed x-axis of plotting route deviation to time
